+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/hooks/useActionCenterTabs.tsx
@@ -5,7 +5,7 @@ import { DiffStatus } from "~/types/api";
 
 export enum ActionCenterTabHash {
   ATTENTION_REQUIRED = "attention-required",
-  RECENT_ACTIVITY = "recent-activity",
+  ADDED = "added",
   IGNORED = "ignored",
 }
 
@@ -25,11 +25,11 @@ const useActionCenterTabs = (systemId?: string) => {
         hash: ActionCenterTabHash.ATTENTION_REQUIRED,
       },
       {
-        label: "Recent activity",
+        label: "Added",
         params: {
           diff_status: [DiffStatus.MONITORED],
         },
-        hash: ActionCenterTabHash.RECENT_ACTIVITY,
+        hash: ActionCenterTabHash.ADDED,
       },
       {
         label: "Ignored",